added error cases to my put, get, and delete routes

+++ server.js
@@ -23,11 +23,15 @@ app
       order: 'id ASC'
     })
     .then( (photos) => {
-      var galleryOfPhotos = [];
-      photos.forEach(function (element) {
-        galleryOfPhotos.push(element.dataValues);
-      });
-      res.render('index', { entries: galleryOfPhotos });
+      if(photos) {
+        var galleryOfPhotos = [];
+        photos.forEach(function (element) {
+          galleryOfPhotos.push(element.dataValues);
+        });
+        res.render('index', { entries: galleryOfPhotos });
+      } else {
+        return next('There are no pictures in the gallery.');
+      }
     });
   })
   .get('/gallery', function (req, res) {
@@ -35,11 +39,15 @@ app
       order: 'id ASC'
     })
     .then( (photos) => {
-      var galleryOfPhotos = [];
-      photos.forEach(function (element) {
-        galleryOfPhotos.push(element.dataValues);
-      });
-      res.render('index', { entries: galleryOfPhotos });
+      if(photos) {
+        var galleryOfPhotos = [];
+        photos.forEach(function (element) {
+          galleryOfPhotos.push(element.dataValues);
+        });
+        res.render('index', { entries: galleryOfPhotos });
+      } else {
+        return next('There are no pictures in the gallery.');
+      }
     });
   })
   .get('/gallery/:id', function (req, res, next) {
@@ -53,11 +61,16 @@ app
           id: { ne: req.params.id }
         }
       });
-      Promise.all([getPhoto, getThreePhotos]).then( (results) => {
-        res.render('get-gallery', {
-          photo: results[0],
-          entries: results[1]
-        });
+      Promise.all([getPhoto, getThreePhotos])
+      .then( (results) => {
+        if(results[0]){
+          res.render('get-gallery', {
+            photo: results[0],
+            entries: results[1]
+          });
+        } else {
+          return next('There is no Gallery Photo of ID: ' + req.params.id + '.');
+        }
       });
     } else {
         res.send('Cannot GET ' + req.params.id);
@@ -67,7 +80,11 @@ app
     if(!isNaN(parseInt(req.params.id))){
       Photo.findById(req.params.id)
       .then( (photo) => {
-        res.render('gallery-edit', photo.dataValues);
+        if(photo){
+          res.render('gallery-edit', photo.dataValues);
+        } else {
+          return next('There is no Gallery Photo of ID: ' + req.params.id + '.');
+        }
       });
     } else {
         res.send('Cannot GET ' + req.params.id);
@@ -80,23 +97,34 @@ app
     });
   })
   .post('/gallery/:id', function (req, res, next) {
-    Photo.create( { url: req.body.url, author: req.body.author, description: req.body.description} )
-    .then( (photo) => {
-      res.render('add-gallery', photo.dataValues);
-    });
+    if(req.params.id === 'new'){
+      Photo.create( { url: req.body.url, author: req.body.author, description: req.body.description} )
+      .then( (photo) => {
+        res.render('add-gallery', photo.dataValues);
+      });
+    } else {
+      res.send('Cannot POST to ' + '/gallery/' + req.params.id);
+    }
   })
   .post('/gallery/:id/edit', function (req, res, next) {
     if(!isNaN(parseInt(req.params.id))){
-      Photo.update( { url: req.body.url, author: req.body.author, description: req.body.description }, {
-        where: {
-          id: req.params.id
+      Photo.findById(req.params.id)
+      .then( (photo) => {
+        if(photo) { // if photo exists, update the photo
+          Photo.update( { url: req.body.url, author: req.body.author, description: req.body.description }, {
+            where: {
+              id: req.params.id
+            }
+          })
+          .then( () => {
+            Photo.findById(req.params.id)
+            .then( (photo) => {
+              res.render('update-gallery', photo.dataValues);
+            });
+          });
+        } else { // if photo doesn't exist
+          return next('There is no Gallery Photo of ID: ' + req.params.id + '.');
         }
-      })
-      .then( () => {
-        Photo.findById(req.params.id)
-        .then( (photo) => {
-          res.render('update-gallery', photo.dataValues);
-        });
       });
     } else {
       res.send('Cannot POST to ' + '/gallery/' + req.params.id);
@@ -104,16 +132,23 @@ app
   })
   .put('/gallery/:id', function (req, res, next) {
     if(!isNaN(parseInt(req.params.id))){
-      Photo.update( { url: req.body.url, author: req.body.author, description: req.body.description }, {
-        where: {
-          id: req.params.id
+      Photo.findById(req.params.id)
+      .then( (photo) => {
+        if(photo) { // if photo exists, update the photo
+          Photo.update( { url: req.body.url, author: req.body.author, description: req.body.description }, {
+            where: {
+              id: req.params.id
+            }
+          })
+          .then( () => {
+            Photo.findById(req.params.id)
+            .then( (photo) => {
+              res.render('update-gallery', photo.dataValues);
+            });
+          });
+        } else { // if photo doesn't exist
+          return next('There is no Gallery Photo of ID: ' + req.params.id + '.');
         }
-      })
-      .then( () => {
-        Photo.findById(req.params.id)
-        .then( (photo) => {
-          res.render('update-gallery', photo.dataValues);
-        });
       });
     } else {
       res.send('Cannot PUT ' + req.params.id);
@@ -123,12 +158,16 @@ app
     if(!isNaN(parseInt(req.params.id))){
       Photo.findById(req.params.id)
       .then( (photo) => {
-        res.render('gallery-delete', photo.dataValues);
-        Photo.destroy({
-          where: {
-            id: req.params.id
-          }
-        });
+        if(photo) {
+          res.render('gallery-delete', photo.dataValues);
+          Photo.destroy({
+            where: {
+              id: req.params.id
+            }
+          });
+        } else {
+          return next('There is no Gallery Photo of ID: ' + req.params.id + '.');
+        }
       });
     } else {
       res.send('Cannot DELETE ' + req.params.id);
